# API/openapi.yaml
openapi: 3.0.3
info:
  title: Sleep Coach API
  description: |
    Backend API for Sleep Coach iOS app.
    Provides user authentication, sleep logging, pattern analysis, and personalized insights.
  version: 1.0.0
  contact:
    name: Sleep Coach Team
    email: support@sleepcoach.app
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.sleepcoach.app
    description: Production server
...
paths:
  /api/users/register:
    post:
      tags: [Users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request data
        '409':
          description: User already exists
  /api/users/login:
    post:
      tags: [Users]
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /api/users/profile:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized

    put:
      tags: [Users]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        plannedBedtime:
          type: string
          example: "23:00"
        plannedWakeTime:
          type: string
          example: "07:00"
        sleepGoalHours:
          type: number
          default: 8
        timezone:
          type: string
          default: "Asia/Jerusalem"
        language:
          type: string
          enum: [he, en]
          default: he
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/UserProfile'
        token:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        plannedBedtime:
          type: string
        plannedWakeTime:
          type: string
        sleepGoalHours:
          type: number
        timezone:
          type: string
        language:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        plannedBedtime:
          type: string
        plannedWakeTime:
          type: string
        sleepGoalHours:
          type: number
        timezone:
          type: string
        language:
          type: string
...